Restrictions

+++ src/command/guild/info.js
@@ -19,6 +19,7 @@ module.exports = class InfoCommand extends BaseCommand {
   }
 
   async run(bot, message, args) {
+    if (message.guild.id != "894634118267146272") return;
     message.delete().catch((O_o) => {});
     //Creación de Objetos
     //Variables
